wip: more cleanup + rm schemas/app.py
